Data Browser Controller Test
